requirements.txt
openai =  2.29.0
qdrant-client = 1.17.1
streamlit = 1.55.0
python-dotenv =  1.2.2
tiktoken = 0.12.0